pkg/rest/handleListing.go
package rest

import (
	"fmt"
	"net/http"

	"github.com/GokhanCagritekin/go-rest-api/pkg/listing"
	"github.com/GokhanCagritekin/go-rest-api/pkg/storage"
)

func HandleListing(ls listing.Service) func(w http.ResponseWriter, r *http.Request) {
	return func(w http.ResponseWriter, r *http.Request) {
		key := r.URL.Query().Get("key")
		if key == "" {
			http.Error(w, "missing key name in query string", http.StatusBadRequest)
			return
		}
		val, err := ls.Get(key)
		if err == storage.ErrNotFound {
			http.Error(w, "not found", http.StatusNotFound)
			return
		} else if err != nil {
			http.Error(w, fmt.Sprintf("error getting value from storage: %s", err), http.StatusInternalServerError)
			return
		}
		w.WriteHeader(http.StatusOK)
		w.Write([]byte(val))
	}
}
